+++ tests/Rules/UrlTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace PHPCommons\Validator\Tests\Rule;
+namespace PHPCommons\Validator\Tests\Rules;
 
-use PHPCommons\Validator\Rule\Url;
+use PHPCommons\Validator\Rules\Url;
 use PHPCommons\Validator\Tests\ResultPair;
 use PHPUnit\Framework\TestCase;
 use function count;
